@@ -1,6 +1,6 @@
 {
   "name": "prettier",
-  "version": "1.10.2",
+  "version": "1.12.1",
   "description": "Prettier is an opinionated code formatter",
   "bin": {
     "prettier": "./bin/prettier.js"
@@ -14,7 +14,7 @@
     "node": ">=4"
   },
   "dependencies": {
-    "@babel/code-frame": "7.0.0-beta.35",
+    "@babel/code-frame": "7.0.0-beta.40",
     "@glimmer/syntax": "0.30.3",
     "babylon": "7.0.0-beta.34",
     "camelcase": "4.1.0",
@@ -24,38 +24,38 @@
     "dashify": "0.2.2",
     "dedent": "0.7.0",
     "diff": "3.2.0",
-    "editorconfig": "0.14.2",
+    "editorconfig": "0.15.0",
     "editorconfig-to-prettier": "0.0.6",
     "emoji-regex": "6.5.1",
     "escape-string-regexp": "1.0.5",
     "esutils": "2.0.2",
+    "find-parent-dir": "0.3.0",
     "find-project-root": "1.1.1",
-    "flow-parser": "0.59.0",
+    "flow-parser": "0.70",
     "get-stream": "3.0.0",
     "globby": "6.1.0",
-    "graphql": "0.12.3",
-    "gray-matter": "3.1.1",
+    "graphql": "0.13.2",
+    "html-tag-names": "1.1.2",
     "ignore": "3.3.7",
-    "jest-docblock": "21.3.0-beta.11",
+    "jest-docblock": "22.2.2",
     "json-stable-stringify": "1.0.1",
     "leven": "2.1.0",
+    "lodash.uniqby": "4.7.0",
     "mem": "1.1.0",
     "minimatch": "3.0.4",
     "minimist": "1.2.0",
     "parse5": "3.0.3",
-    "postcss-less": "1.1.3",
+    "postcss-less": "1.1.5",
     "postcss-media-query-parser": "0.2.3",
-    "postcss-scss": "1.0.2",
+    "postcss-scss": "1.0.5",
     "postcss-selector-parser": "2.2.3",
-    "postcss-values-parser": "1.3.1",
-    "read-pkg-up": "3.0.0",
-    "remark-frontmatter": "1.1.0",
+    "postcss-values-parser": "1.5.0",
     "remark-parse": "5.0.0",
     "resolve": "1.5.0",
     "semver": "5.4.1",
     "string-width": "2.1.1",
-    "typescript": "2.7.0-insiders.20171214",
-    "typescript-eslint-parser": "11.0.0",
+    "typescript": "2.9.0-dev.20180421",
+    "typescript-eslint-parser": "eslint/typescript-eslint-parser#2960b002746c01fb9cb15bb5f4c1e7e925c6519a",
     "unicode-regex": "1.0.1",
     "unified": "6.1.6"
   },
@@ -64,15 +64,15 @@
     "babel-preset-es2015": "6.24.1",
     "codecov": "2.2.0",
     "cross-env": "5.0.5",
-    "eslint": "4.1.1",
+    "eslint": "4.18.2",
     "eslint-config-prettier": "2.9.0",
     "eslint-friendly-formatter": "3.0.0",
-    "eslint-plugin-import": "2.6.1",
-    "eslint-plugin-prettier": "2.4.0",
-    "eslint-plugin-react": "7.1.0",
+    "eslint-plugin-import": "2.9.0",
+    "eslint-plugin-prettier": "2.6.0",
+    "eslint-plugin-react": "7.7.0",
     "jest": "21.1.0",
     "mkdirp": "0.5.1",
-    "prettier": "1.10.2",
+    "prettier": "1.12.1",
     "prettylint": "1.0.0",
     "rimraf": "2.6.2",
     "rollup": "0.47.6",
@@ -85,9 +85,8 @@
     "shelljs": "0.8.1",
     "snapshot-diff": "0.2.2",
     "strip-ansi": "4.0.0",
-    "sw-toolbox": "3.6.0",
     "tempy": "0.2.1",
-    "uglify-es": "3.0.28",
+    "uglify-es": "3.3.9",
     "webpack": "2.6.1"
   },
   "scripts": {
