@@ -3,24 +3,23 @@
 const path = require("path");
 const camelCase = require("camelcase");
 const dashify = require("dashify");
-const minimist = require("minimist");
 const fs = require("fs");
 const globby = require("globby");
-const ignore = require("ignore");
 const chalk = require("chalk");
 const readline = require("readline");
 const leven = require("leven");
+const stringify = require("json-stable-stringify");
 
+const minimist = require("./minimist");
 const prettier = require("../../index");
-const cleanAST = require("../common/clean-ast").cleanAST;
+const createIgnorer = require("../common/create-ignorer");
 const errors = require("../common/errors");
-const resolver = require("../config/resolve-config");
 const constant = require("./constant");
+const coreOptions = require("../main/core-options");
 const optionsModule = require("../main/options");
 const optionsNormalizer = require("../main/options-normalizer");
 const thirdParty = require("../common/third-party");
-const getSupportInfo = require("../common/support").getSupportInfo;
-const util = require("../common/util");
+const arrayify = require("../utils/arrayify");
 
 const OPTION_USAGE_THRESHOLD = 25;
 const CHOICE_USAGE_MARGIN = 3;
@@ -78,13 +77,32 @@ function handleError(context, filename, error) {
   process.exitCode = 2;
 }
 
-function logResolvedConfigPathOrDie(context, filePath) {
-  const configFile = resolver.resolveConfigFile.sync(filePath);
+function logResolvedConfigPathOrDie(context) {
+  const configFile = prettier.resolveConfigFile.sync(
+    context.argv["find-config-path"]
+  );
   if (configFile) {
     context.logger.log(path.relative(process.cwd(), configFile));
   } else {
     process.exit(1);
   }
+}
+
+function logFileInfoOrDie(context) {
+  const options = {
+    ignorePath: context.argv["ignore-path"],
+    withNodeModules: context.argv["with-node-modules"],
+    plugins: context.argv["plugin"],
+    pluginSearchDirs: context.argv["plugin-search-dir"]
+  };
+  context.logger.log(
+    prettier.format(
+      stringify(prettier.getFileInfo.sync(context.argv["file-info"], options)),
+      {
+        parser: "json"
+      }
+    )
+  );
 }
 
 function writeOutput(result, options) {
@@ -127,14 +145,12 @@ function format(context, input, opt) {
         "prettier(input) !== prettier(prettier(input))\n" + diff(pp, pppp)
       );
     } else {
-      const normalizedOpts = optionsModule.normalize(opt);
-      const ast = cleanAST(
-        prettier.__debug.parse(input, opt).ast,
-        normalizedOpts
+      const stringify = obj => JSON.stringify(obj, null, 2);
+      const ast = stringify(
+        prettier.__debug.parse(input, opt, /* massage */ true).ast
       );
-      const past = cleanAST(
-        prettier.__debug.parse(pp, opt).ast,
-        normalizedOpts
+      const past = stringify(
+        prettier.__debug.parse(pp, opt, /* massage */ true).ast
       );
 
       if (ast !== past) {
@@ -172,7 +188,7 @@ function getOptionsOrDie(context, filePath) {
         : `resolve config from '${filePath}'`
     );
 
-    const options = resolver.resolveConfig.sync(filePath, {
+    const options = prettier.resolveConfig.sync(filePath, {
       editorconfig: context.argv["editorconfig"],
       config: context.argv["config"]
     });
@@ -228,11 +244,7 @@ function parseArgsToOptions(context, overrideDefaults) {
         Object.assign({
           string: minimistOptions.string,
           boolean: minimistOptions.boolean,
-          default: Object.assign(
-            {},
-            cliifyOptions(context.apiDefaultOptions, apiDetailedOptionMap),
-            cliifyOptions(overrideDefaults, apiDetailedOptionMap)
-          )
+          default: cliifyOptions(overrideDefaults, apiDetailedOptionMap)
         })
       ),
       context.detailedOptions,
@@ -263,7 +275,7 @@ function formatStdin(context) {
     ? path.resolve(process.cwd(), context.argv["stdin-filepath"])
     : process.cwd();
 
-  const ignorer = createIgnorer(context);
+  const ignorer = createIgnorerFromContextOrDie(context);
   const relativeFilepath = path.relative(process.cwd(), filepath);
 
   thirdParty.getStream(process.stdin).then(input => {
@@ -286,26 +298,20 @@ function formatStdin(context) {
   });
 }
 
-function createIgnorer(context) {
-  const ignoreFilePath = path.resolve(context.argv["ignore-path"]);
-  let ignoreText = "";
-
+function createIgnorerFromContextOrDie(context) {
   try {
-    ignoreText = fs.readFileSync(ignoreFilePath, "utf8");
-  } catch (readError) {
-    if (readError.code !== "ENOENT") {
-      context.logger.error(
-        `Unable to read ${ignoreFilePath}: ` + readError.message
-      );
-      process.exit(2);
-    }
+    return createIgnorer(
+      context.argv["ignore-path"],
+      context.argv["with-node-modules"]
+    );
+  } catch (e) {
+    context.logger.error(e.message);
+    process.exit(2);
   }
-
-  return ignore().add(ignoreText);
 }
 
 function eachFilename(context, patterns, callback) {
-  const ignoreNodeModules = context.argv["with-node-modules"] === false;
+  const ignoreNodeModules = context.argv["with-node-modules"] !== true;
   if (ignoreNodeModules) {
     patterns = patterns.concat(["!**/node_modules/**", "!./node_modules/**"]);
   }
@@ -337,13 +343,15 @@ function eachFilename(context, patterns, callback) {
 function formatFiles(context) {
   // The ignorer will be used to filter file paths after the glob is checked,
   // before any files are actually written
-  const ignorer = createIgnorer(context);
+  const ignorer = createIgnorerFromContextOrDie(context);
 
   eachFilename(context, context.filePatterns, (filename, options) => {
     const fileIgnored = ignorer.filter([filename]).length === 0;
     if (
       fileIgnored &&
-      (context.argv["write"] || context.argv["list-different"])
+      (context.argv["debug-check"] ||
+        context.argv["write"] ||
+        context.argv["list-different"])
     ) {
       return;
     }
@@ -613,7 +621,16 @@ function createDetailedUsage(context, optionName) {
       ? `\n\nDefault: ${createDefaultValueDisplay(optionDefaultValue)}`
       : "";
 
-  return `${header}${description}${choices}${defaults}`;
+  const pluginDefaults =
+    option.pluginDefaults && Object.keys(option.pluginDefaults).length
+      ? `\nPlugin defaults:${Object.keys(option.pluginDefaults).map(
+          key =>
+            `\n* ${key}: ${createDefaultValueDisplay(
+              option.pluginDefaults[key]
+            )}`
+        )}`
+      : "";
+  return `${header}${description}${choices}${defaults}${pluginDefaults}`;
 }
 
 function getOptionDefaultValue(context, optionName) {
@@ -706,7 +723,7 @@ function createLogger(logLevel) {
 }
 
 function normalizeDetailedOption(name, option) {
-  return Object.assign({ category: constant.CATEGORY_OTHER }, option, {
+  return Object.assign({ category: coreOptions.CATEGORY_OTHER }, option, {
     choices:
       option.choices &&
       option.choices.map(choice => {
@@ -743,6 +760,12 @@ function createMinimistOptions(detailedOptions) {
       .map(option => option.name),
     default: detailedOptions
       .filter(option => !option.deprecated)
+      .filter(
+        option =>
+          !option.forwardToApi ||
+          option.name === "plugin" ||
+          option.name === "plugin-search-dir"
+      )
       .filter(option => option.default !== undefined)
       .reduce(
         (current, option) =>
@@ -774,7 +797,7 @@ function createDetailedOptionMap(supportOptions) {
     const newOption = Object.assign({}, option, {
       name: option.cliName || dashify(option.name),
       description: option.cliDescription || option.description,
-      category: option.cliCategory || constant.CATEGORY_FORMAT,
+      category: option.cliCategory || coreOptions.CATEGORY_FORMAT,
       forwardToApi: option.name
     });
 
@@ -805,11 +828,15 @@ function createContext(args) {
   const context = { args };
 
   updateContextArgv(context);
-  normalizeContextArgv(context, ["loglevel", "plugin"]);
+  normalizeContextArgv(context, ["loglevel", "plugin", "plugin-search-dir"]);
 
   context.logger = createLogger(context.argv["loglevel"]);
 
-  updateContextArgv(context, context.argv["plugin"]);
+  updateContextArgv(
+    context,
+    context.argv["plugin"],
+    context.argv["plugin-search-dir"]
+  );
 
   return context;
 }
@@ -819,19 +846,20 @@ function initContext(context) {
   normalizeContextArgv(context);
 }
 
-function updateContextOptions(context, plugins) {
-  const supportOptions = getSupportInfo(null, {
+function updateContextOptions(context, plugins, pluginSearchDirs) {
+  const supportOptions = prettier.getSupportInfo(null, {
     showDeprecated: true,
     showUnreleased: true,
     showInternal: true,
-    plugins
+    plugins,
+    pluginSearchDirs
   }).options;
 
   const detailedOptionMap = normalizeDetailedOptionMap(
     Object.assign({}, createDetailedOptionMap(supportOptions), constant.options)
   );
 
-  const detailedOptions = util.arrayify(detailedOptionMap, "name");
+  const detailedOptions = arrayify(detailedOptionMap, "name");
 
   const apiDefaultOptions = supportOptions
     .filter(optionInfo => !optionInfo.deprecated)
@@ -847,12 +875,12 @@ function updateContextOptions(context, plugins) {
   context.apiDefaultOptions = apiDefaultOptions;
 }
 
-function pushContextPlugins(context, plugins) {
+function pushContextPlugins(context, plugins, pluginSearchDirs) {
   context._supportOptions = context.supportOptions;
   context._detailedOptions = context.detailedOptions;
   context._detailedOptionMap = context.detailedOptionMap;
   context._apiDefaultOptions = context.apiDefaultOptions;
-  updateContextOptions(context, plugins);
+  updateContextOptions(context, plugins, pluginSearchDirs);
 }
 
 function popContextPlugins(context) {
@@ -862,8 +890,8 @@ function popContextPlugins(context) {
   context.apiDefaultOptions = context._apiDefaultOptions;
 }
 
-function updateContextArgv(context, plugins) {
-  pushContextPlugins(context, plugins);
+function updateContextArgv(context, plugins, pluginSearchDirs) {
+  pushContextPlugins(context, plugins, pluginSearchDirs);
 
   const minimistOptions = createMinimistOptions(context.detailedOptions);
   const argv = minimist(context.args, minimistOptions);
@@ -896,5 +924,6 @@ module.exports = {
   formatStdin,
   initContext,
   logResolvedConfigPathOrDie,
+  logFileInfoOrDie,
   normalizeDetailedOptionMap
 };
