@@ -1,10 +1,30 @@
 "use strict";
 
+const uniqBy = require("lodash.uniqby");
+const fs = require("fs");
+const globby = require("globby");
+const path = require("path");
 const resolve = require("resolve");
-const readPkgUp = require("read-pkg-up");
+const thirdParty = require("./third-party");
 
-function loadPlugins(plugins) {
-  plugins = plugins || [];
+function loadPlugins(plugins, pluginSearchDirs) {
+  if (!plugins) {
+    plugins = [];
+  }
+
+  if (!pluginSearchDirs) {
+    pluginSearchDirs = [];
+  }
+  // unless pluginSearchDirs are provided, auto-load plugins from node_modules that are parent to Prettier
+  if (!pluginSearchDirs.length) {
+    const autoLoadDir = thirdParty.findParentDir(
+      thirdParty.findParentDir(__dirname, "prettier"),
+      "node_modules"
+    );
+    if (autoLoadDir) {
+      pluginSearchDirs = [autoLoadDir];
+    }
+  }
 
   const internalPlugins = [
     require("../language-js"),
@@ -16,45 +36,74 @@ function loadPlugins(plugins) {
     require("../language-vue")
   ];
 
-  const externalPlugins = plugins
-    .concat(
-      getPluginsFromPackage(
-        readPkgUp.sync({
-          normalize: false
-        }).pkg
-      )
-    )
-    .map(plugin => {
-      if (typeof plugin !== "string") {
-        return plugin;
+  const externalManualLoadPluginInfos = plugins.map(pluginName => {
+    let requirePath;
+    try {
+      // try local files
+      requirePath = resolve.sync(path.resolve(process.cwd(), pluginName));
+    } catch (e) {
+      // try node modules
+      requirePath = resolve.sync(pluginName, { basedir: process.cwd() });
+    }
+    return {
+      name: pluginName,
+      requirePath
+    };
+  });
+
+  const externalAutoLoadPluginInfos = pluginSearchDirs
+    .map(pluginSearchDir => {
+      const resolvedPluginSearchDir = path.resolve(
+        process.cwd(),
+        pluginSearchDir
+      );
+
+      if (!isDirectory(resolvedPluginSearchDir)) {
+        throw new Error(
+          `${pluginSearchDir} does not exist or is not a directory`
+        );
       }
 
-      const pluginPath = resolve.sync(plugin, { basedir: process.cwd() });
-      return eval("require")(pluginPath);
-    });
+      const nodeModulesDir = path.resolve(
+        resolvedPluginSearchDir,
+        "node_modules"
+      );
 
-  return deduplicate(internalPlugins.concat(externalPlugins));
-}
+      return findPluginsInNodeModules(nodeModulesDir).map(pluginName => ({
+        name: pluginName,
+        requirePath: resolve.sync(pluginName, {
+          basedir: resolvedPluginSearchDir
+        })
+      }));
+    })
+    .reduce((a, b) => a.concat(b), []);
 
-function getPluginsFromPackage(pkg) {
-  if (!pkg) {
-    return [];
-  }
-  const deps = Object.assign({}, pkg.dependencies, pkg.devDependencies);
-  return Object.keys(deps).filter(
-    dep =>
-      dep.startsWith("prettier-plugin-") || dep.startsWith("@prettier/plugin-")
+  const externalPlugins = uniqBy(
+    externalManualLoadPluginInfos.concat(externalAutoLoadPluginInfos),
+    "requirePath"
+  ).map(externalPluginInfo =>
+    Object.assign(
+      { name: externalPluginInfo.name },
+      eval("require")(externalPluginInfo.requirePath)
+    )
   );
+
+  return internalPlugins.concat(externalPlugins);
 }
 
-function deduplicate(items) {
-  const uniqItems = [];
-  for (const item of items) {
-    if (uniqItems.indexOf(item) < 0) {
-      uniqItems.push(item);
-    }
+function findPluginsInNodeModules(nodeModulesDir) {
+  const pluginPackageJsonPaths = globby.sync(
+    ["prettier-plugin-*/package.json", "@prettier/plugin-*/package.json"],
+    { cwd: nodeModulesDir }
+  );
+  return pluginPackageJsonPaths.map(path.dirname);
+}
+
+function isDirectory(dir) {
+  try {
+    return fs.statSync(dir).isDirectory();
+  } catch (e) {
+    return false;
   }
-  return uniqItems;
 }
-
 module.exports = loadPlugins;
