@@ -2,8 +2,10 @@
 
 const getStream = require("get-stream");
 const cosmiconfig = require("cosmiconfig");
+const findParentDir = require("find-parent-dir").sync;
 
 module.exports = {
   getStream,
-  cosmiconfig
+  cosmiconfig,
+  findParentDir
 };
