@@ -462,6 +462,7 @@ function printDocToString(doc, options) {
                   // Trim whitespace at the end of line
                   while (
                     out.length > 0 &&
+                    typeof out[out.length - 1] === "string" &&
                     out[out.length - 1].match(/^[^\S\n]*$/)
                   ) {
                     out.pop();
@@ -469,6 +470,7 @@ function printDocToString(doc, options) {
 
                   if (
                     out.length &&
+                    typeof out[out.length - 1] === "string" &&
                     (options.parser !== "markdown" ||
                       // preserve markdown's `break` node (two trailing spaces)
                       !/\S {2}$/.test(out[out.length - 1]))
@@ -493,12 +495,20 @@ function printDocToString(doc, options) {
 
   const cursorPlaceholderIndex = out.indexOf(cursor.placeholder);
   if (cursorPlaceholderIndex !== -1) {
+    const otherCursorPlaceholderIndex = out.indexOf(
+      cursor.placeholder,
+      cursorPlaceholderIndex + 1
+    );
     const beforeCursor = out.slice(0, cursorPlaceholderIndex).join("");
-    const afterCursor = out.slice(cursorPlaceholderIndex + 1).join("");
+    const aroundCursor = out
+      .slice(cursorPlaceholderIndex + 1, otherCursorPlaceholderIndex)
+      .join("");
+    const afterCursor = out.slice(otherCursorPlaceholderIndex + 1).join("");
 
     return {
-      formatted: beforeCursor + afterCursor,
-      cursor: beforeCursor.length
+      formatted: beforeCursor + aroundCursor + afterCursor,
+      cursorNodeStart: beforeCursor.length,
+      cursorNodeText: aroundCursor
     };
   }
 
