@@ -2,7 +2,7 @@
 
 const jsOptions = require("../language-js/options");
 
-// format based on https://github.com/prettier/prettier/blob/master/src/common/support.js
+// format based on https://github.com/prettier/prettier/blob/master/src/main/core-options.js
 module.exports = {
   singleQuote: jsOptions.singleQuote
 };
