@@ -1,26 +1,13 @@
 "use strict";
 
 const createError = require("../common/parser-create-error");
-const grayMatter = require("gray-matter");
+const parseFrontmatter = require("../utils/front-matter");
 
-function parseSelector(selector) {
-  // If there's a comment inside of a selector, the parser tries to parse
-  // the content of the comment as selectors which turns it into complete
-  // garbage. Better to print the whole selector as-is and not try to parse
-  // and reformat it.
-  if (selector.match(/\/\/|\/\*/)) {
-    return {
-      type: "selector-comment",
-      value: selector
-    };
-  }
-  const selectorParser = require("postcss-selector-parser");
-  let result;
-  selectorParser(result_ => {
-    result = result_;
-  }).process(selector);
-  return addTypePrefix(result, "selector-");
-}
+// utils
+const utils = require("./utils");
+
+const isSCSS = utils.isSCSS;
+const isSCSSNestedPropertyNode = utils.isSCSSNestedPropertyNode;
 
 function parseValueNodes(nodes) {
   let parenGroup = {
@@ -54,8 +41,6 @@ function parseValueNodes(nodes) {
 
     if (isUnquotedDataURLCall) {
       node.group.groups = [stringifyGroup(node)];
-
-      return node;
     }
 
     if (node.type === "paren" && node.value === "(") {
@@ -195,19 +180,79 @@ function parseNestedValue(node) {
 
 function parseValue(value) {
   const valueParser = require("postcss-values-parser");
-  const result = valueParser(value, { loose: true }).parse();
+
+  let result = null;
+
+  try {
+    result = valueParser(value, { loose: true }).parse();
+  } catch (e) {
+    return {
+      type: "value-unknown",
+      value: value
+    };
+  }
+
   const parsedResult = parseNestedValue(result);
+
   return addTypePrefix(parsedResult, "value-");
 }
 
-function parseMediaQuery(value) {
-  const mediaParser = require("postcss-media-query-parser").default;
-  const result = addMissingType(mediaParser(value));
-  return addTypePrefix(result, "media-");
+function parseSelector(selector) {
+  // If there's a comment inside of a selector, the parser tries to parse
+  // the content of the comment as selectors which turns it into complete
+  // garbage. Better to print the whole selector as-is and not try to parse
+  // and reformat it.
+  if (selector.match(/\/\/|\/\*/)) {
+    return {
+      type: "selector-unknown",
+      value: selector.replace(/^ +/, "").replace(/ +$/, "")
+    };
+  }
+
+  const selectorParser = require("postcss-selector-parser");
+
+  let result = null;
+
+  try {
+    selectorParser(result_ => {
+      result = result_;
+    }).process(selector);
+  } catch (e) {
+    // Fail silently. It's better to print it as is than to try and parse it
+    // Note: A common failure is for SCSS nested properties. `background:
+    // none { color: red; }` is parsed as a NestedDeclaration by
+    // postcss-scss, while `background: { color: red; }` is parsed as a Rule
+    // with a selector ending with a colon. See:
+    // https://github.com/postcss/postcss-scss/issues/39
+    return {
+      type: "selector-unknown",
+      value: selector
+    };
+  }
+
+  return addTypePrefix(result, "selector-");
 }
 
-const DEFAULT_SCSS_DIRECTIVE = "!default";
-const GLOBAL_SCSS_DIRECTIVE = "!global";
+function parseMediaQuery(params) {
+  const mediaParser = require("postcss-media-query-parser").default;
+
+  let result = null;
+
+  try {
+    result = mediaParser(params);
+  } catch (e) {
+    // Ignore bad media queries
+    return {
+      type: "selector-unknown",
+      value: params
+    };
+  }
+
+  return addTypePrefix(addMissingType(result), "media-");
+}
+
+const DEFAULT_SCSS_DIRECTIVE = /(\s*?)(!default).*$/;
+const GLOBAL_SCSS_DIRECTIVE = /(\s*?)(!global).*$/;
 
 function parseNestedCSS(node) {
   if (node && typeof node === "object") {
@@ -217,147 +262,209 @@ function parseNestedCSS(node) {
       parseNestedCSS(node[key]);
     }
 
-    if (typeof node.selector === "string" && node.selector.trim().length > 0) {
-      const selector = node.raws.selector
-        ? node.raws.selector.raw
-        : node.selector;
+    if (!node.type) {
+      return node;
+    }
 
+    if (!node.raws) {
+      node.raws = {};
+    }
+
+    let selector = "";
+
+    if (typeof node.selector === "string") {
+      selector =
+        node.raws.selector && node.raws.selector.raw
+          ? node.raws.selector.raw
+          : node.selector;
+
+      if (node.raws.between && node.raws.between.trim().length > 0) {
+        selector += node.raws.between;
+      }
+
+      node.raws.selector = selector;
+    }
+
+    let value = "";
+
+    if (typeof node.value === "string") {
+      value =
+        node.raws.value && node.raws.value.raw
+          ? node.raws.value.raw
+          : node.value;
+
+      value = value.trim();
+
+      node.raws.value = selector;
+    }
+
+    let params = "";
+
+    if (typeof node.params === "string") {
+      params =
+        node.raws.params && node.raws.params.raw
+          ? node.raws.params.raw
+          : node.params;
+
+      if (node.raws.afterName && node.raws.afterName.trim().length > 0) {
+        params = node.raws.afterName + params;
+      }
+
+      if (node.raws.between && node.raws.between.trim().length > 0) {
+        params = params + node.raws.between;
+      }
+
+      params = params.trim();
+
+      node.raws.params = params;
+    }
+
+    // Ignore LESS mixin declaration
+    if (selector.trim().length > 0) {
       if (selector.startsWith("@") && selector.endsWith(":")) {
         return node;
       }
 
-      try {
-        node.selector = parseSelector(selector);
-      } catch (e) {
-        // Fail silently. It's better to print it as is than to try and parse it
-        // Note: A common failure is for SCSS nested properties. `background:
-        // none { color: red; }` is parsed as a NestedDeclaration by
-        // postcss-scss, while `background: { color: red; }` is parsed as a Rule
-        // with a selector ending with a colon. See:
-        // https://github.com/postcss/postcss-scss/issues/39
-        node.selector = {
-          type: "selector-root-invalid",
-          value: selector
-        };
-      }
+      // Ignore LESS mixins
+      if (node.mixin) {
+        node.selector = parseValue(selector);
 
-      return node;
-    }
-
-    if (
-      node.type &&
-      node.type !== "css-comment-yaml" &&
-      typeof node.value === "string" &&
-      node.value.trim().length > 0
-    ) {
-      try {
-        if (node.value.endsWith(DEFAULT_SCSS_DIRECTIVE)) {
-          node.default = true;
-          node.value = node.value.slice(0, -DEFAULT_SCSS_DIRECTIVE.length);
-        }
-
-        if (node.value.endsWith(GLOBAL_SCSS_DIRECTIVE)) {
-          node.global = true;
-          node.value = node.value.slice(0, -GLOBAL_SCSS_DIRECTIVE.length);
-        }
-
-        node.value = parseValue(node.value);
-      } catch (e) {
-        throw createError(
-          "(postcss-values-parser) " + e.toString(),
-          node.source
-        );
-      }
-
-      return node;
-    }
-
-    if (
-      node.type === "css-atrule" &&
-      typeof node.params === "string" &&
-      node.params.trim().length > 0
-    ) {
-      if (
-        node.name === "charset" ||
-        node.name.toLowerCase() === "counter-style" ||
-        node.name.toLowerCase().endsWith("keyframes") ||
-        node.name.toLowerCase() === "page" ||
-        node.name.toLowerCase() === "font-feature-values"
-      ) {
         return node;
       }
 
-      if (node.name === "warn" || node.name === "error") {
+      // Check on SCSS nested property
+      if (isSCSSNestedPropertyNode(node)) {
+        node.isSCSSNesterProperty = true;
+      }
+
+      node.selector = parseSelector(selector);
+
+      return node;
+    }
+
+    if (node.type !== "css-comment-yaml" && value.length > 0) {
+      const defaultSCSSDirectiveIndex = value.match(DEFAULT_SCSS_DIRECTIVE);
+
+      if (defaultSCSSDirectiveIndex) {
+        value = value.substring(0, defaultSCSSDirectiveIndex.index);
+        node.scssDefault = true;
+
+        if (defaultSCSSDirectiveIndex[0].trim() !== "!default") {
+          node.raws.scssDefault = defaultSCSSDirectiveIndex[0];
+        }
+      }
+
+      const globalSCSSDirectiveIndex = value.match(GLOBAL_SCSS_DIRECTIVE);
+
+      if (globalSCSSDirectiveIndex) {
+        value = value.substring(0, globalSCSSDirectiveIndex.index);
+        node.scssGlobal = true;
+
+        if (globalSCSSDirectiveIndex[0].trim() !== "!global") {
+          node.raws.scssGlobal = globalSCSSDirectiveIndex[0];
+        }
+      }
+
+      if (value.startsWith("progid:")) {
+        return {
+          type: "value-unknown",
+          value: value
+        };
+      }
+
+      node.value = parseValue(value);
+    }
+
+    if (node.type === "css-atrule" && params.length > 0) {
+      const name = node.name;
+      const lowercasedName = node.name.toLowerCase();
+
+      if (name === "warn" || name === "error") {
         node.params = {
           type: "media-unknown",
-          value: node.params
+          value: params
         };
 
         return node;
       }
 
-      if (node.name === "extend") {
-        node.selector = parseSelector(node.params);
+      if (name === "extend" || name === "nest") {
+        node.selector = parseSelector(params);
         delete node.params;
 
         return node;
       }
 
-      if (node.name === "at-root") {
-        if (/^\(\s*(without|with)\s*:[\s\S]+\)$/.test(node.params)) {
-          node.params = parseMediaQuery(node.params);
+      if (name === "at-root") {
+        if (/^\(\s*(without|with)\s*:[\s\S]+\)$/.test(params)) {
+          node.params = parseValue(params);
         } else {
-          node.selector = parseSelector(node.params);
+          node.selector = parseSelector(params);
           delete node.params;
         }
 
         return node;
       }
 
+      if (lowercasedName === "import") {
+        node.params = parseValue(params);
+        return node;
+      }
+
       if (
-        node.name === "if" ||
-        node.name === "else" ||
-        node.name === "for" ||
-        node.name === "each" ||
-        node.name === "while" ||
-        node.name === "debug" ||
-        node.name === "mixin" ||
-        node.name === "include" ||
-        node.name === "function" ||
-        node.name === "return"
+        [
+          "namespace",
+          "supports",
+          "if",
+          "else",
+          "for",
+          "each",
+          "while",
+          "debug",
+          "mixin",
+          "include",
+          "function",
+          "return",
+          "define-mixin",
+          "add-mixin"
+        ].indexOf(name) !== -1
       ) {
         // Remove unnecessary spaces in SCSS variable arguments
-        node.params = node.params.replace(/\s+\.\.\./, "...");
+        params = params.replace(/(\$\S+?)\s+?\.\.\./, "$1...");
         // Remove unnecessary spaces before SCSS control, mixin and function directives
-        node.params = node.params.replace(/^(\S+)\s+\(/, "$1(");
+        params = params.replace(/^(?!if)(\S+)\s+\(/, "$1(");
 
-        node.value = parseValue(node.params);
+        node.value = parseValue(params);
         delete node.params;
 
         return node;
       }
 
-      if (node.params.includes("#{")) {
-        // Workaround for media at rule with scss interpolation
-        return {
-          type: "media-unknown",
-          value: node.params
-        };
-      }
-
-      if (/^custom-selector$/i.test(node.name)) {
-        const customSelector = node.params.match(/:--\S+?\s+/)[0].trim();
+      if (name === "custom-selector") {
+        const customSelector = params.match(/:--\S+?\s+/)[0].trim();
 
         node.customSelector = customSelector;
-        node.selector = parseSelector(
-          node.params.substring(customSelector.length)
-        );
+        node.selector = parseSelector(params.substring(customSelector.length));
         delete node.params;
 
         return node;
       }
 
-      node.params = parseMediaQuery(node.params);
+      if (["media", "custom-media"].indexOf(lowercasedName) !== -1) {
+        if (params.includes("#{")) {
+          // Workaround for media at rule with scss interpolation
+          return {
+            type: "media-unknown",
+            value: params
+          };
+        }
+
+        node.params = parseMediaQuery(params);
+
+        return node;
+      }
+
+      node.params = params;
 
       return node;
     }
@@ -366,8 +473,13 @@ function parseNestedCSS(node) {
   return node;
 }
 
-function parseWithParser(parser, text, frontMatter) {
+function parseWithParser(parser, text) {
+  const parsed = parseFrontmatter(text);
+  const frontmatter = parsed.frontmatter;
+  text = parsed.content;
+
   let result;
+
   try {
     result = parser.parse(text);
   } catch (e) {
@@ -377,20 +489,20 @@ function parseWithParser(parser, text, frontMatter) {
     throw createError("(postcss) " + e.name + " " + e.reason, { start: e });
   }
 
-  if (Object.keys(frontMatter.data).length > 0) {
+  result = parseNestedCSS(addTypePrefix(result, "css-"));
+
+  if (frontmatter) {
     result.nodes.unshift({
-      type: "comment-yaml",
-      value: grayMatter.stringify("", frontMatter.data).replace(/\s$/, "")
+      type: "frontmatter",
+      value: frontmatter
     });
   }
 
-  const prefixedResult = addTypePrefix(result, "css-");
-  const parsedResult = parseNestedCSS(prefixedResult);
-  return parsedResult;
+  return result;
 }
 
-function requireParser(isSCSS) {
-  if (isSCSS) {
+function requireParser(isSCSSParser) {
+  if (isSCSSParser) {
     return require("postcss-scss");
   }
 
@@ -407,32 +519,20 @@ function requireParser(isSCSS) {
   return require("postcss-less");
 }
 
-const IS_POSSIBLY_SCSS = /(\w\s*: [^}:]+|#){|@import[^\n]+(url|,)/;
-
 function parse(text, parsers, opts) {
   const hasExplicitParserChoice =
     opts.parser === "less" || opts.parser === "scss";
-
-  const isSCSS = hasExplicitParserChoice
-    ? opts.parser === "scss"
-    : IS_POSSIBLY_SCSS.test(text);
-
-  const frontMatter = grayMatter(text);
-  const normalizedText = frontMatter.content;
+  const isSCSSParser = isSCSS(opts.parser, text);
 
   try {
-    return parseWithParser(requireParser(isSCSS), normalizedText, frontMatter);
+    return parseWithParser(requireParser(isSCSSParser), text);
   } catch (originalError) {
     if (hasExplicitParserChoice) {
       throw originalError;
     }
 
     try {
-      return parseWithParser(
-        requireParser(!isSCSS),
-        normalizedText,
-        frontMatter
-      );
+      return parseWithParser(requireParser(!isSCSSParser), text);
     } catch (_secondError) {
       throw originalError;
     }
