@@ -1,7 +1,8 @@
 "use strict";
 
 const embed = require("./embed");
-const util = require("../common/util");
+const clean = require("./clean");
+const privateUtil = require("../common/util");
 const docBuilders = require("../doc").builders;
 const concat = docBuilders.concat;
 const join = docBuilders.join;
@@ -10,7 +11,6 @@ const line = docBuilders.line;
 const softline = docBuilders.softline;
 const group = docBuilders.group;
 const indent = docBuilders.indent;
-// const ifBreak = docBuilders.ifBreak;
 
 // http://w3c.github.io/html/single-page.html#void-elements
 const voidTags = {
@@ -56,10 +56,10 @@ function genericPrint(path, options, print) {
       const selfClose = voidTags[n.name] ? ">" : " />";
       const children = printChildren(path, print);
 
-      const hasNewline = util.hasNewlineInRange(
+      const hasNewline = privateUtil.hasNewlineInRange(
         options.originalText,
-        util.locStart(n),
-        util.locEnd(n)
+        options.locStart(n),
+        options.locEnd(n)
       );
 
       return group(
@@ -117,6 +117,7 @@ function printChildren(path, print) {
 
 module.exports = {
   print: genericPrint,
+  massageAstNode: clean,
   embed,
-  hasPrettierIgnore: util.hasIgnoreComment
+  hasPrettierIgnore: privateUtil.hasIgnoreComment
 };
