@@ -1,9 +1,68 @@
 "use strict";
 
-const printer = require("./printer-estree");
+const estreePrinter = require("./printer-estree");
+const estreeJsonPrinter = require("./printer-estree-json");
+const hasPragma = require("./pragma").hasPragma;
+const options = require("./options");
+const privateUtil = require("../common/util");
 
 // Based on:
 // https://github.com/github/linguist/blob/master/lib/linguist/languages.yml
+
+const locStart = function(node) {
+  // Handle nodes with decorators. They should start at the first decorator
+  if (
+    node.declaration &&
+    node.declaration.decorators &&
+    node.declaration.decorators.length > 0
+  ) {
+    return locStart(node.declaration.decorators[0]);
+  }
+  if (node.decorators && node.decorators.length > 0) {
+    return locStart(node.decorators[0]);
+  }
+
+  if (node.__location) {
+    return node.__location.startOffset;
+  }
+  if (node.range) {
+    return node.range[0];
+  }
+  if (typeof node.start === "number") {
+    return node.start;
+  }
+  if (node.loc) {
+    return node.loc.start;
+  }
+  return null;
+};
+
+const locEnd = function(node) {
+  const endNode = node.nodes && privateUtil.getLast(node.nodes);
+  if (endNode && node.source && !node.source.end) {
+    node = endNode;
+  }
+
+  let loc;
+  if (node.range) {
+    loc = node.range[1];
+  } else if (typeof node.end === "number") {
+    loc = node.end;
+  }
+
+  if (node.__location) {
+    return node.__location.endOffset;
+  }
+  if (node.typeAnnotation) {
+    return Math.max(loc, locEnd(node.typeAnnotation));
+  }
+
+  if (node.loc && !loc) {
+    return node.loc.end;
+  }
+
+  return loc;
+};
 
 const languages = [
   {
@@ -70,6 +129,20 @@ const languages = [
     vscodeLanguageIds: ["typescript", "typescriptreact"]
   },
   {
+    name: "JSON.stringify",
+    since: "1.13.0",
+    parsers: ["json-stringify"],
+    group: "JavaScript",
+    tmScope: "source.json",
+    aceMode: "json",
+    codemirrorMode: "javascript",
+    codemirrorMimeType: "application/json",
+    extensions: [], // .json file defaults to json instead of json-stringify
+    filenames: ["package.json", "package-lock.json"],
+    linguistLanguageId: 174,
+    vscodeLanguageIds: ["json"]
+  },
+  {
     name: "JSON",
     since: "1.5.0",
     parsers: ["json"],
@@ -78,17 +151,10 @@ const languages = [
     aceMode: "json",
     codemirrorMode: "javascript",
     codemirrorMimeType: "application/json",
-    extensions: [
-      ".json",
-      ".json5",
-      ".geojson",
-      ".JSON-tmLanguage",
-      ".topojson"
-    ],
+    extensions: [".json", ".geojson", ".JSON-tmLanguage", ".topojson"],
     filenames: [
       ".arcconfig",
       ".jshintrc",
-      ".babelrc",
       ".eslintrc",
       ".prettierrc",
       "composer.lock",
@@ -96,6 +162,20 @@ const languages = [
     ],
     linguistLanguageId: 174,
     vscodeLanguageIds: ["json", "jsonc"]
+  },
+  {
+    name: "JSON5",
+    since: "1.13.0",
+    parsers: ["json5"],
+    group: "JavaScript",
+    tmScope: "source.json",
+    aceMode: "json",
+    codemirrorMode: "javascript",
+    codemirrorMimeType: "application/json",
+    extensions: [".json5"],
+    filenames: [".babelrc"],
+    linguistLanguageId: 175,
+    vscodeLanguageIds: ["json5"]
   }
 ];
 
@@ -103,24 +183,46 @@ const typescript = {
   get parse() {
     return eval("require")("./parser-typescript");
   },
-  astFormat: "estree"
+  astFormat: "estree",
+  hasPragma,
+  locStart,
+  locEnd
 };
 
 const babylon = {
   get parse() {
     return eval("require")("./parser-babylon");
   },
-  astFormat: "estree"
+  astFormat: "estree",
+  hasPragma,
+  locStart,
+  locEnd
 };
 
 const parsers = {
   babylon,
-  json: babylon,
+  json: Object.assign({}, babylon, {
+    hasPragma() {
+      return true;
+    }
+  }),
+  json5: babylon,
+  "json-stringify": {
+    get parse() {
+      return eval("require")("./parser-json-stringify");
+    },
+    astFormat: "estree-json",
+    locStart,
+    locEnd
+  },
   flow: {
     get parse() {
       return eval("require")("./parser-flow");
     },
-    astFormat: "estree"
+    astFormat: "estree",
+    hasPragma,
+    locStart,
+    locEnd
   },
   "typescript-eslint": typescript,
   // TODO: Delete this in 2.0
@@ -128,11 +230,13 @@ const parsers = {
 };
 
 const printers = {
-  estree: printer
+  estree: estreePrinter,
+  "estree-json": estreeJsonPrinter
 };
 
 module.exports = {
   languages,
+  options,
   parsers,
   printers
 };
