@@ -12,7 +12,8 @@ function parse(text /*, parsers, opts*/) {
   const ast = flowParser.parse(text, {
     esproposal_class_instance_fields: true,
     esproposal_class_static_fields: true,
-    esproposal_export_star_as: true
+    esproposal_export_star_as: true,
+    esproposal_optional_chaining: true
   });
 
   if (ast.errors.length > 0) {
