@@ -1,6 +1,8 @@
 "use strict";
 
 const printer = require("./printer-markdown");
+const options = require("./options");
+const pragma = require("./pragma");
 
 // Based on:
 // https://github.com/github/linguist/blob/master/lib/linguist/languages.yml
@@ -37,7 +39,10 @@ const remark = {
   get parse() {
     return eval("require")("./parser-markdown");
   },
-  astFormat: "mdast"
+  astFormat: "mdast",
+  hasPragma: pragma.hasPragma,
+  locStart: node => node.position.start.offset,
+  locEnd: node => node.position.end.offset
 };
 
 const parsers = {
@@ -52,6 +57,7 @@ const printers = {
 
 module.exports = {
   languages,
+  options,
   parsers,
   printers
 };
