@@ -1,7 +1,9 @@
 "use strict";
 
-const util = require("../common/util");
+const docUtils = require("../doc/doc-utils");
+const privateUtil = require("../common/util");
 const embed = require("./embed");
+const pragma = require("./pragma");
 const doc = require("../doc");
 const docBuilders = doc.builders;
 const concat = docBuilders.concat;
@@ -11,9 +13,9 @@ const hardline = docBuilders.hardline;
 const softline = docBuilders.softline;
 const fill = docBuilders.fill;
 const align = docBuilders.align;
+const indent = docBuilders.indent;
 const group = docBuilders.group;
 const printDocToString = doc.printer.printDocToString;
-const printerOptions = require("./options");
 
 const SINGLE_LINE_NODE_TYPES = ["heading", "tableCell", "link"];
 
@@ -47,7 +49,7 @@ function genericPrint(path, options, print) {
 
   if (shouldRemainTheSameContent(path)) {
     return concat(
-      util
+      privateUtil
         .splitText(
           options.originalText.slice(
             node.position.start.offset,
@@ -58,17 +60,19 @@ function genericPrint(path, options, print) {
           node =>
             node.type === "word"
               ? node.value
-              : node.value === "" ? "" : printLine(path, node.value, options)
+              : node.value === ""
+                ? ""
+                : printLine(path, node.value, options)
         )
     );
   }
 
   switch (node.type) {
     case "root":
-      return concat([
-        normalizeDoc(printChildren(path, options, print)),
-        hardline
-      ]);
+      if (node.children.length === 0) {
+        return "";
+      }
+      return concat([normalizeDoc(printRoot(path, options, print)), hardline]);
     case "paragraph":
       return printChildren(path, options, print, {
         postprocessor: fill
@@ -81,8 +85,8 @@ function genericPrint(path, options, print) {
         .replace(
           new RegExp(
             [
-              `(^|[${util.punctuationCharRange}])(_+)`,
-              `(_+)([${util.punctuationCharRange}]|$)`
+              `(^|[${privateUtil.punctuationCharRange}])(_+)`,
+              `(_+)([${privateUtil.punctuationCharRange}]|$)`
             ].join("|"),
             "g"
           ),
@@ -114,8 +118,8 @@ function genericPrint(path, options, print) {
         (prevNode &&
           prevNode.type === "sentence" &&
           prevNode.children.length > 0 &&
-          util.getLast(prevNode.children).type === "word" &&
-          !util.getLast(prevNode.children).hasTrailingPunctuation) ||
+          privateUtil.getLast(prevNode.children).type === "word" &&
+          !privateUtil.getLast(prevNode.children).hasTrailingPunctuation) ||
         (nextNode &&
           nextNode.type === "sentence" &&
           nextNode.children.length > 0 &&
@@ -130,7 +134,7 @@ function genericPrint(path, options, print) {
     case "delete":
       return concat(["~~", printChildren(path, options, print), "~~"]);
     case "inlineCode": {
-      const backtickCount = util.getMaxContinuousCount(node.value, "`");
+      const backtickCount = privateUtil.getMaxContinuousCount(node.value, "`");
       const style = backtickCount === 1 ? "``" : "`";
       const gap = backtickCount ? " " : "";
       return concat([style, gap, node.value, gap, style]);
@@ -191,7 +195,10 @@ function genericPrint(path, options, print) {
       // fenced code block
       const styleUnit = options.__inJsTemplate ? "~" : "`";
       const style = styleUnit.repeat(
-        Math.max(3, util.getMaxContinuousCount(node.value, styleUnit) + 1)
+        Math.max(
+          3,
+          privateUtil.getMaxContinuousCount(node.value, styleUnit) + 1
+        )
       );
       return concat([
         style,
@@ -202,16 +209,16 @@ function genericPrint(path, options, print) {
         style
       ]);
     }
-    case "yaml":
-      return concat(["---", hardline, node.value, hardline, "---"]);
-    case "toml":
-      return concat(["+++", hardline, node.value, hardline, "+++"]);
+    case "frontmatter":
+      return node.value;
     case "html": {
       const parentNode = path.getParentNode();
-      return parentNode.type === "root" &&
-        util.getLast(parentNode.children) === node
-        ? node.value.trimRight()
-        : node.value;
+      return replaceNewlinesWithHardlines(
+        parentNode.type === "root" &&
+        privateUtil.getLast(parentNode.children) === node
+          ? node.value.trimRight()
+          : node.value
+      );
     }
     case "list": {
       const nthSiblingIndex = getNthListSiblingIndex(
@@ -231,12 +238,7 @@ function genericPrint(path, options, print) {
 
       return printChildren(path, options, print, {
         processor: (childPath, index) => {
-          const prefix = node.ordered
-            ? (index === 0
-                ? node.start
-                : isGitDiffFriendlyOrderedList ? 1 : node.start + index) +
-              (nthSiblingIndex % 2 === 0 ? ". " : ") ")
-            : nthSiblingIndex % 2 === 0 ? "* " : "- ";
+          const prefix = getPrefix();
           return concat([
             prefix,
             align(
@@ -244,6 +246,27 @@ function genericPrint(path, options, print) {
               printListItem(childPath, options, print, prefix)
             )
           ]);
+
+          function getPrefix() {
+            const rawPrefix = node.ordered
+              ? (index === 0
+                  ? node.start
+                  : isGitDiffFriendlyOrderedList
+                    ? 1
+                    : node.start + index) +
+                (nthSiblingIndex % 2 === 0 ? ". " : ") ")
+              : nthSiblingIndex % 2 === 0
+                ? "- "
+                : "* ";
+
+            // do not print trailing spaces for empty list item since it might be treated as `break` node
+            // by [doc-printer](https://github.com/prettier/prettier/blob/1.10.2/src/doc/doc-printer.js#L395-L405),
+            // we don't want to preserve unnecessary trailing spaces.
+            const listItem = childPath.getValue();
+            return listItem.children.length
+              ? alignListPrefix(rawPrefix, options)
+              : rawPrefix;
+          }
         }
       });
     }
@@ -256,7 +279,7 @@ function genericPrint(path, options, print) {
         path.getParentNode(counter),
         path.getParentNode(counter + 1)
       );
-      return nthSiblingIndex % 2 === 0 ? "---" : "***";
+      return nthSiblingIndex % 2 === 0 ? "***" : "---";
     }
     case "linkReference":
       return concat([
@@ -265,7 +288,9 @@ function genericPrint(path, options, print) {
         "]",
         node.referenceType === "full"
           ? concat(["[", node.identifier, "]"])
-          : node.referenceType === "collapsed" ? "[]" : ""
+          : node.referenceType === "collapsed"
+            ? "[]"
+            : ""
       ]);
     case "imageReference":
       switch (node.referenceType) {
@@ -279,14 +304,23 @@ function genericPrint(path, options, print) {
             node.referenceType === "collapsed" ? "[]" : ""
           ]);
       }
-    case "definition":
-      return concat([
-        "[",
-        node.identifier,
-        "]: ",
-        printUrl(node.url),
-        printTitle(node.title, options)
-      ]);
+    case "definition": {
+      const lineOrSpace = options.proseWrap === "always" ? line : " ";
+      return group(
+        concat([
+          concat(["[", node.identifier, "]:"]),
+          indent(
+            concat([
+              lineOrSpace,
+              printUrl(node.url),
+              node.title === null
+                ? ""
+                : concat([lineOrSpace, printTitle(node.title, options, false)])
+            ])
+          )
+        ])
+      );
+    }
     case "footnote":
       return concat(["[^", printChildren(path, options, print), "]"]);
     case "footnoteReference":
@@ -343,12 +377,27 @@ function printListItem(path, options, print, listPrefix) {
         }
 
         const alignment = " ".repeat(
-          clamp(options.tabWidth - listPrefix.length, 0, 3) // 4 will cause indented codeblock
+          clamp(options.tabWidth - listPrefix.length, 0, 3) // 4+ will cause indented code block
         );
         return concat([alignment, align(alignment, childPath.call(print))]);
       }
     })
   ]);
+}
+
+function alignListPrefix(prefix, options) {
+  const additionalSpaces = getAdditionalSpaces();
+  return (
+    prefix +
+    " ".repeat(
+      additionalSpaces >= 4 ? 0 : additionalSpaces // 4+ will cause indented code block
+    )
+  );
+
+  function getAdditionalSpaces() {
+    const restSpaces = prefix.length % options.tabWidth;
+    return restSpaces === 0 ? 0 : options.tabWidth - restSpaces;
+  }
 }
 
 function getNthListSiblingIndex(node, parentNode) {
@@ -357,6 +406,10 @@ function getNthListSiblingIndex(node, parentNode) {
     parentNode,
     siblingNode => siblingNode.ordered === node.ordered
   );
+}
+
+function replaceNewlinesWithHardlines(str) {
+  return join(hardline, str.split("\n"));
 }
 
 function getNthSiblingIndex(node, parentNode, condition) {
@@ -406,8 +459,12 @@ function printLine(path, value, options) {
     options.proseWrap === "always" &&
     !getAncestorNode(path, SINGLE_LINE_NODE_TYPES);
   return value !== ""
-    ? isBreakable ? line : " "
-    : isBreakable ? softline : "";
+    ? isBreakable
+      ? line
+      : " "
+    : isBreakable
+      ? softline
+      : "";
 }
 
 function printTable(path, options, print) {
@@ -429,7 +486,7 @@ function printTable(path, options, print) {
   const columnMaxWidths = contents.reduce(
     (currentWidths, rowContents) =>
       currentWidths.map((width, columnIndex) =>
-        Math.max(width, util.getStringWidth(rowContents[columnIndex]))
+        Math.max(width, privateUtil.getStringWidth(rowContents[columnIndex]))
       ),
     contents[0].map(() => 3) // minimum width = 3 (---, :--, :-:, --:)
   );
@@ -483,19 +540,81 @@ function printTable(path, options, print) {
   }
 
   function alignLeft(text, width) {
-    return concat([text, " ".repeat(width - util.getStringWidth(text))]);
+    return concat([text, " ".repeat(width - privateUtil.getStringWidth(text))]);
   }
 
   function alignRight(text, width) {
-    return concat([" ".repeat(width - util.getStringWidth(text)), text]);
+    return concat([" ".repeat(width - privateUtil.getStringWidth(text)), text]);
   }
 
   function alignCenter(text, width) {
-    const spaces = width - util.getStringWidth(text);
+    const spaces = width - privateUtil.getStringWidth(text);
     const left = Math.floor(spaces / 2);
     const right = spaces - left;
     return concat([" ".repeat(left), text, " ".repeat(right)]);
   }
+}
+
+function printRoot(path, options, print) {
+  /** @typedef {{ index: number, offset: number }} IgnorePosition */
+  /** @type {Array<{start: IgnorePosition, end: IgnorePosition}>} */
+  const ignoreRanges = [];
+
+  /** @type {IgnorePosition | null} */
+  let ignoreStart = null;
+
+  const children = path.getValue().children;
+  children.forEach((childNode, index) => {
+    switch (isPrettierIgnore(childNode)) {
+      case "start":
+        if (ignoreStart === null) {
+          ignoreStart = { index, offset: childNode.position.end.offset };
+        }
+        break;
+      case "end":
+        if (ignoreStart !== null) {
+          ignoreRanges.push({
+            start: ignoreStart,
+            end: { index, offset: childNode.position.start.offset }
+          });
+          ignoreStart = null;
+        }
+        break;
+      default:
+        // do nothing
+        break;
+    }
+  });
+
+  return printChildren(path, options, print, {
+    processor: (childPath, index) => {
+      if (ignoreRanges.length !== 0) {
+        const ignoreRange = ignoreRanges[0];
+
+        if (index === ignoreRange.start.index) {
+          return concat([
+            children[ignoreRange.start.index].value,
+            options.originalText.slice(
+              ignoreRange.start.offset,
+              ignoreRange.end.offset
+            ),
+            children[ignoreRange.end.index].value
+          ]);
+        }
+
+        if (ignoreRange.start.index < index && index < ignoreRange.end.index) {
+          return false;
+        }
+
+        if (index === ignoreRange.end.index) {
+          ignoreRanges.shift();
+          return false;
+        }
+      }
+
+      return childPath.call(print);
+    }
+  });
 }
 
 function printChildren(path, options, print, events) {
@@ -507,28 +626,15 @@ function printChildren(path, options, print, events) {
   const node = path.getValue();
   const parts = [];
 
-  let counter = 0;
   let lastChildNode;
-  let prettierIgnore = false;
 
   path.map((childPath, index) => {
     const childNode = childPath.getValue();
 
-    const result = prettierIgnore
-      ? options.originalText.slice(
-          childNode.position.start.offset,
-          childNode.position.end.offset
-        )
-      : processor(childPath, index);
-
-    prettierIgnore = false;
-
+    const result = processor(childPath, index);
     if (result !== false) {
-      prettierIgnore = isPrettierIgnore(childNode);
-
       const data = {
         parts,
-        index: counter++,
         prevNode: lastChildNode,
         parentNode: node,
         options
@@ -558,10 +664,15 @@ function printChildren(path, options, print, events) {
   return postprocessor(parts);
 }
 
+/** @return {false | 'next' | 'start' | 'end'} */
 function isPrettierIgnore(node) {
-  return (
-    node.type === "html" && /^<!--\s*prettier-ignore\s*-->$/.test(node.value)
+  if (node.type !== "html") {
+    return false;
+  }
+  const match = node.value.match(
+    /^<!--\s*prettier-ignore(?:-(start|end))?\s*-->$/
   );
+  return match === null ? false : match[1] ? match[1] : "next";
 }
 
 function shouldNotPrePrintHardline(node, data) {
@@ -586,7 +697,7 @@ function shouldPrePrintDoubleHardline(node, data) {
   const isPrevNodeLooseListItem =
     data.prevNode && data.prevNode.type === "listItem" && data.prevNode.loose;
 
-  const isPrevNodePrettierIgnore = isPrettierIgnore(data.prevNode);
+  const isPrevNodePrettierIgnore = isPrettierIgnore(data.prevNode) === "next";
 
   return (
     isPrevNodeLooseListItem ||
@@ -617,7 +728,7 @@ function shouldRemainTheSameContent(path) {
 }
 
 function normalizeDoc(doc) {
-  return util.mapDoc(doc, currentDoc => {
+  return docUtils.mapDoc(doc, currentDoc => {
     if (!currentDoc.parts) {
       return currentDoc;
     }
@@ -649,12 +760,19 @@ function printUrl(url, dangerousCharOrChars) {
     : url;
 }
 
-function printTitle(title, options) {
+function printTitle(title, options, printSpace) {
+  if (printSpace == null) {
+    printSpace = true;
+  }
+
   if (!title) {
     return "";
   }
+  if (printSpace) {
+    return " " + printTitle(title, options, false);
+  }
   if (title.includes('"') && title.includes("'") && !title.includes(")")) {
-    return ` (${title})`; // avoid escaped quotes
+    return `(${title})`; // avoid escaped quotes
   }
   // faster than using RegExps: https://jsperf.com/performance-of-match-vs-split
   const singleCount = title.split("'").length - 1;
@@ -662,14 +780,18 @@ function printTitle(title, options) {
   const quote =
     singleCount > doubleCount
       ? '"'
-      : doubleCount > singleCount ? "'" : options.singleQuote ? "'" : '"';
+      : doubleCount > singleCount
+        ? "'"
+        : options.singleQuote
+          ? "'"
+          : '"';
   title = title.replace(new RegExp(`(${quote})`, "g"), "\\$1");
-  return ` ${quote}${title}${quote}`;
+  return `${quote}${title}${quote}`;
 }
 
 function normalizeParts(parts) {
   return parts.reduce((current, part) => {
-    const lastPart = util.getLast(current);
+    const lastPart = privateUtil.getLast(current);
 
     if (typeof lastPart === "string" && typeof part === "string") {
       current.splice(-1, 1, lastPart + part);
@@ -685,21 +807,47 @@ function clamp(value, min, max) {
   return value < min ? min : value > max ? max : value;
 }
 
-function clean(ast, newObj) {
-  // for markdown codeblock
+function clean(ast, newObj, parent) {
+  delete newObj.position;
+
+  // for codeblock
   if (ast.type === "code") {
     delete newObj.value;
   }
-  // for markdown whitespace: "\n" and " " are considered the same
+  // for whitespace: "\n" and " " are considered the same
   if (ast.type === "whitespace" && ast.value === "\n") {
     newObj.value = " ";
   }
+  // for insert pragma
+  if (
+    parent &&
+    parent.type === "root" &&
+    parent.children.length > 0 &&
+    (parent.children[0] === ast ||
+      (parent.children[0].type === "frontmatter" &&
+        parent.children[1] === ast)) &&
+    ast.type === "html" &&
+    pragma.startWithPragma(ast.value)
+  ) {
+    return null;
+  }
+}
+
+function hasPrettierIgnore(path) {
+  const index = +path.getName();
+
+  if (index === 0) {
+    return false;
+  }
+
+  const prevNode = path.getParentNode().children[index - 1];
+  return isPrettierIgnore(prevNode) === "next";
 }
 
 module.exports = {
-  options: printerOptions,
   print: genericPrint,
   embed,
   massageAstNode: clean,
-  hasPrettierIgnore: util.hasIgnoreComment
+  hasPrettierIgnore,
+  insertPragma: pragma.insertPragma
 };
