@@ -9,20 +9,21 @@ const indent = docBuilders.indent;
 const lineSuffix = docBuilders.lineSuffix;
 const join = docBuilders.join;
 const cursor = docBuilders.cursor;
-const util = require("../common/util");
+const privateUtil = require("../common/util");
+const sharedUtil = require("../common/util-shared");
 const childNodesCacheKey = Symbol("child-nodes");
-const locStart = util.locStart;
-const locEnd = util.locEnd;
-const getNextNonSpaceNonCommentCharacter =
-  util.getNextNonSpaceNonCommentCharacter;
-const getNextNonSpaceNonCommentCharacterIndex =
-  util.getNextNonSpaceNonCommentCharacterIndex;
 
-function getSortedChildNodes(node, text, options, resultArray) {
+const addLeadingComment = sharedUtil.addLeadingComment;
+const addTrailingComment = sharedUtil.addTrailingComment;
+const addDanglingComment = sharedUtil.addDanglingComment;
+
+function getSortedChildNodes(node, options, resultArray) {
   if (!node) {
     return;
   }
   const printer = options.printer;
+  const locStart = options.locStart;
+  const locEnd = options.locEnd;
 
   if (resultArray) {
     if (node && printer.canAttachComment && printer.canAttachComment(node)) {
@@ -72,7 +73,7 @@ function getSortedChildNodes(node, text, options, resultArray) {
   }
 
   childNodes.forEach(childNode => {
-    getSortedChildNodes(childNode, text, options, resultArray);
+    getSortedChildNodes(childNode, options, resultArray);
   });
 
   return resultArray;
@@ -81,8 +82,10 @@ function getSortedChildNodes(node, text, options, resultArray) {
 // As efficiently as possible, decorate the comment object with
 // .precedingNode, .enclosingNode, and/or .followingNode properties, at
 // least one of which is guaranteed to be defined.
-function decorateComment(node, comment, text, options) {
-  const childNodes = getSortedChildNodes(node, text, options);
+function decorateComment(node, comment, options) {
+  const locStart = options.locStart;
+  const locEnd = options.locEnd;
+  const childNodes = getSortedChildNodes(node, options);
   let precedingNode;
   let followingNode;
   // Time to dust off the old binary search robes and wizard hat.
@@ -99,7 +102,7 @@ function decorateComment(node, comment, text, options) {
       // The comment is completely contained by this child node.
       comment.enclosingNode = child;
 
-      decorateComment(child, comment, text, options);
+      decorateComment(child, comment, options);
       return; // Abandon the binary search at this level.
     }
 
@@ -134,17 +137,23 @@ function decorateComment(node, comment, text, options) {
     comment.enclosingNode.type === "TemplateLiteral"
   ) {
     const quasis = comment.enclosingNode.quasis;
-    const commentIndex = findExpressionIndexForComment(quasis, comment);
+    const commentIndex = findExpressionIndexForComment(
+      quasis,
+      comment,
+      options
+    );
 
     if (
       precedingNode &&
-      findExpressionIndexForComment(quasis, precedingNode) !== commentIndex
+      findExpressionIndexForComment(quasis, precedingNode, options) !==
+        commentIndex
     ) {
       precedingNode = null;
     }
     if (
       followingNode &&
-      findExpressionIndexForComment(quasis, followingNode) !== commentIndex
+      findExpressionIndexForComment(quasis, followingNode, options) !==
+        commentIndex
     ) {
       followingNode = null;
     }
@@ -165,64 +174,44 @@ function attach(comments, ast, text, options) {
   }
 
   const tiesToBreak = [];
+  const locStart = options.locStart;
+  const locEnd = options.locEnd;
 
   comments.forEach((comment, i) => {
-    if (options.parser === "json" && locStart(comment) - locStart(ast) <= 0) {
+    if (
+      (options.parser === "json" || options.parser === "json5") &&
+      locStart(comment) - locStart(ast) <= 0
+    ) {
       addLeadingComment(ast, comment);
       return;
     }
 
-    decorateComment(ast, comment, text, options);
+    decorateComment(ast, comment, options);
 
     const precedingNode = comment.precedingNode;
     const enclosingNode = comment.enclosingNode;
     const followingNode = comment.followingNode;
 
+    const pluginHandleOwnLineComment =
+      options.printer.handleComments && options.printer.handleComments.ownLine
+        ? options.printer.handleComments.ownLine
+        : () => false;
+    const pluginHandleEndOfLineComment =
+      options.printer.handleComments && options.printer.handleComments.endOfLine
+        ? options.printer.handleComments.endOfLine
+        : () => false;
+    const pluginHandleRemainingComment =
+      options.printer.handleComments && options.printer.handleComments.remaining
+        ? options.printer.handleComments.remaining
+        : () => false;
+
     const isLastComment = comments.length - 1 === i;
 
-    if (util.hasNewline(text, locStart(comment), { backwards: true })) {
+    if (privateUtil.hasNewline(text, locStart(comment), { backwards: true })) {
       // If a comment exists on its own line, prefer a leading comment.
       // We also need to check if it's the first line of the file.
       if (
-        handleLastFunctionArgComments(
-          text,
-          precedingNode,
-          enclosingNode,
-          followingNode,
-          comment
-        ) ||
-        handleMemberExpressionComments(enclosingNode, followingNode, comment) ||
-        handleIfStatementComments(
-          text,
-          precedingNode,
-          enclosingNode,
-          followingNode,
-          comment
-        ) ||
-        handleTryStatementComments(enclosingNode, followingNode, comment) ||
-        handleClassComments(
-          enclosingNode,
-          precedingNode,
-          followingNode,
-          comment
-        ) ||
-        handleImportSpecifierComments(enclosingNode, comment) ||
-        handleForComments(enclosingNode, precedingNode, comment) ||
-        handleUnionTypeComments(
-          precedingNode,
-          enclosingNode,
-          followingNode,
-          comment
-        ) ||
-        handleOnlyComments(enclosingNode, ast, comment, isLastComment) ||
-        handleImportDeclarationComments(
-          text,
-          enclosingNode,
-          precedingNode,
-          comment
-        ) ||
-        handleAssignmentPatternComments(enclosingNode, comment) ||
-        handleMethodNameComments(text, enclosingNode, precedingNode, comment)
+        pluginHandleOwnLineComment(comment, text, options, ast, isLastComment)
       ) {
         // We're good
       } else if (followingNode) {
@@ -237,43 +226,9 @@ function attach(comments, ast, text, options) {
         /* istanbul ignore next */
         addDanglingComment(ast, comment);
       }
-    } else if (util.hasNewline(text, locEnd(comment))) {
+    } else if (privateUtil.hasNewline(text, locEnd(comment))) {
       if (
-        handleLastFunctionArgComments(
-          text,
-          precedingNode,
-          enclosingNode,
-          followingNode,
-          comment
-        ) ||
-        handleConditionalExpressionComments(
-          enclosingNode,
-          precedingNode,
-          followingNode,
-          comment,
-          text
-        ) ||
-        handleImportSpecifierComments(enclosingNode, comment) ||
-        handleIfStatementComments(
-          text,
-          precedingNode,
-          enclosingNode,
-          followingNode,
-          comment
-        ) ||
-        handleClassComments(
-          enclosingNode,
-          precedingNode,
-          followingNode,
-          comment
-        ) ||
-        handleLabeledStatementComments(enclosingNode, comment) ||
-        handleCallExpressionComments(precedingNode, enclosingNode, comment) ||
-        handlePropertyComments(enclosingNode, comment) ||
-        handleExportNamedDeclarationComments(enclosingNode, comment) ||
-        handleOnlyComments(enclosingNode, ast, comment, isLastComment) ||
-        handleTypeAliasComments(enclosingNode, followingNode, comment) ||
-        handleVariableDeclaratorComments(enclosingNode, followingNode, comment)
+        pluginHandleEndOfLineComment(comment, text, options, ast, isLastComment)
       ) {
         // We're good
       } else if (precedingNode) {
@@ -291,19 +246,7 @@ function attach(comments, ast, text, options) {
       }
     } else {
       if (
-        handleIfStatementComments(
-          text,
-          precedingNode,
-          enclosingNode,
-          followingNode,
-          comment
-        ) ||
-        handleObjectPropertyAssignment(enclosingNode, precedingNode, comment) ||
-        handleCommentInEmptyParens(text, enclosingNode, comment) ||
-        handleMethodNameComments(text, enclosingNode, precedingNode, comment) ||
-        handleOnlyComments(enclosingNode, ast, comment, isLastComment) ||
-        handleCommentAfterArrowParams(text, enclosingNode, comment) ||
-        handleFunctionNameComments(text, enclosingNode, precedingNode, comment)
+        pluginHandleRemainingComment(comment, text, options, ast, isLastComment)
       ) {
         // We're good
       } else if (precedingNode && followingNode) {
@@ -316,7 +259,7 @@ function attach(comments, ast, text, options) {
         if (tieCount > 0) {
           const lastTie = tiesToBreak[tieCount - 1];
           if (lastTie.followingNode !== comment.followingNode) {
-            breakTies(tiesToBreak, text);
+            breakTies(tiesToBreak, text, options);
           }
         }
         tiesToBreak.push(comment);
@@ -334,7 +277,7 @@ function attach(comments, ast, text, options) {
     }
   });
 
-  breakTies(tiesToBreak, text);
+  breakTies(tiesToBreak, text, options);
 
   comments.forEach(comment => {
     // These node references were useful for breaking ties, but we
@@ -346,7 +289,7 @@ function attach(comments, ast, text, options) {
   });
 }
 
-function breakTies(tiesToBreak, text) {
+function breakTies(tiesToBreak, text, options) {
   const tieCount = tiesToBreak.length;
   if (tieCount === 0) {
     return;
@@ -354,7 +297,7 @@ function breakTies(tiesToBreak, text) {
 
   const precedingNode = tiesToBreak[0].precedingNode;
   const followingNode = tiesToBreak[0].followingNode;
-  let gapEndPos = locStart(followingNode);
+  let gapEndPos = options.locStart(followingNode);
 
   // Iterate backwards through tiesToBreak, examining the gaps
   // between the tied comments. In order to qualify as leading, a
@@ -371,9 +314,9 @@ function breakTies(tiesToBreak, text) {
     assert.strictEqual(comment.precedingNode, precedingNode);
     assert.strictEqual(comment.followingNode, followingNode);
 
-    const gap = text.slice(locEnd(comment), gapEndPos).trim();
+    const gap = text.slice(options.locEnd(comment), gapEndPos).trim();
     if (gap === "" || /^\(+$/.test(gap)) {
-      gapEndPos = locStart(comment);
+      gapEndPos = options.locStart(comment);
     } else {
       // The gap string contained something other than whitespace or open
       // parentheses.
@@ -392,540 +335,14 @@ function breakTies(tiesToBreak, text) {
   tiesToBreak.length = 0;
 }
 
-function addCommentHelper(node, comment) {
-  const comments = node.comments || (node.comments = []);
-  comments.push(comment);
-  comment.printed = false;
-
-  // For some reason, TypeScript parses `// x` inside of JSXText as a comment
-  // We already "print" it via the raw text, we don't need to re-print it as a
-  // comment
-  if (node.type === "JSXText") {
-    comment.printed = true;
-  }
-}
-
-function addLeadingComment(node, comment) {
-  comment.leading = true;
-  comment.trailing = false;
-  addCommentHelper(node, comment);
-}
-
-function addDanglingComment(node, comment) {
-  comment.leading = false;
-  comment.trailing = false;
-  addCommentHelper(node, comment);
-}
-
-function addTrailingComment(node, comment) {
-  comment.leading = false;
-  comment.trailing = true;
-  addCommentHelper(node, comment);
-}
-
-function addBlockStatementFirstComment(node, comment) {
-  const body = node.body.filter(n => n.type !== "EmptyStatement");
-  if (body.length === 0) {
-    addDanglingComment(node, comment);
-  } else {
-    addLeadingComment(body[0], comment);
-  }
-}
-
-function addBlockOrNotComment(node, comment) {
-  if (node.type === "BlockStatement") {
-    addBlockStatementFirstComment(node, comment);
-  } else {
-    addLeadingComment(node, comment);
-  }
-}
-
-// There are often comments before the else clause of if statements like
-//
-//   if (1) { ... }
-//   // comment
-//   else { ... }
-//
-// They are being attached as leading comments of the BlockExpression which
-// is not well printed. What we want is to instead move the comment inside
-// of the block and make it leadingComment of the first element of the block
-// or dangling comment of the block if there is nothing inside
-//
-//   if (1) { ... }
-//   else {
-//     // comment
-//     ...
-//   }
-function handleIfStatementComments(
-  text,
-  precedingNode,
-  enclosingNode,
-  followingNode,
-  comment
-) {
-  if (
-    !enclosingNode ||
-    enclosingNode.type !== "IfStatement" ||
-    !followingNode
-  ) {
-    return false;
-  }
-
-  // We unfortunately have no way using the AST or location of nodes to know
-  // if the comment is positioned before the condition parenthesis:
-  //   if (a /* comment */) {}
-  // The only workaround I found is to look at the next character to see if
-  // it is a ).
-  const nextCharacter = getNextNonSpaceNonCommentCharacter(text, comment);
-  if (nextCharacter === ")") {
-    addTrailingComment(precedingNode, comment);
-    return true;
-  }
-
-  if (followingNode.type === "BlockStatement") {
-    addBlockStatementFirstComment(followingNode, comment);
-    return true;
-  }
-
-  if (followingNode.type === "IfStatement") {
-    addBlockOrNotComment(followingNode.consequent, comment);
-    return true;
-  }
-
-  // For comments positioned after the condition parenthesis in an if statement
-  // before the consequent with or without brackets on, such as
-  // if (a) /* comment */ {} or if (a) /* comment */ true,
-  // we look at the next character to see if it is a { or if the following node
-  // is the consequent for the if statement
-  if (nextCharacter === "{" || enclosingNode.consequent === followingNode) {
-    addLeadingComment(followingNode, comment);
-    return true;
-  }
-
-  return false;
-}
-
-// Same as IfStatement but for TryStatement
-function handleTryStatementComments(enclosingNode, followingNode, comment) {
-  if (
-    !enclosingNode ||
-    enclosingNode.type !== "TryStatement" ||
-    !followingNode
-  ) {
-    return false;
-  }
-
-  if (followingNode.type === "BlockStatement") {
-    addBlockStatementFirstComment(followingNode, comment);
-    return true;
-  }
-
-  if (followingNode.type === "TryStatement") {
-    addBlockOrNotComment(followingNode.finalizer, comment);
-    return true;
-  }
-
-  if (followingNode.type === "CatchClause") {
-    addBlockOrNotComment(followingNode.body, comment);
-    return true;
-  }
-
-  return false;
-}
-
-function handleMemberExpressionComments(enclosingNode, followingNode, comment) {
-  if (
-    enclosingNode &&
-    enclosingNode.type === "MemberExpression" &&
-    followingNode &&
-    followingNode.type === "Identifier"
-  ) {
-    addLeadingComment(enclosingNode, comment);
-    return true;
-  }
-
-  return false;
-}
-
-function handleConditionalExpressionComments(
-  enclosingNode,
-  precedingNode,
-  followingNode,
-  comment,
-  text
-) {
-  const isSameLineAsPrecedingNode =
-    precedingNode &&
-    !util.hasNewlineInRange(text, locEnd(precedingNode), locStart(comment));
-
-  if (
-    (!precedingNode || !isSameLineAsPrecedingNode) &&
-    enclosingNode &&
-    enclosingNode.type === "ConditionalExpression" &&
-    followingNode
-  ) {
-    addLeadingComment(followingNode, comment);
-    return true;
-  }
-  return false;
-}
-
-function handleObjectPropertyAssignment(enclosingNode, precedingNode, comment) {
-  if (
-    enclosingNode &&
-    (enclosingNode.type === "ObjectProperty" ||
-      enclosingNode.type === "Property") &&
-    enclosingNode.shorthand &&
-    enclosingNode.key === precedingNode &&
-    enclosingNode.value.type === "AssignmentPattern"
-  ) {
-    addTrailingComment(enclosingNode.value.left, comment);
-    return true;
-  }
-  return false;
-}
-
-function handleClassComments(
-  enclosingNode,
-  precedingNode,
-  followingNode,
-  comment
-) {
-  if (
-    enclosingNode &&
-    (enclosingNode.type === "ClassDeclaration" ||
-      enclosingNode.type === "ClassExpression") &&
-    (enclosingNode.decorators && enclosingNode.decorators.length > 0) &&
-    !(followingNode && followingNode.type === "Decorator")
-  ) {
-    if (!enclosingNode.decorators || enclosingNode.decorators.length === 0) {
-      addLeadingComment(enclosingNode, comment);
-    } else {
-      addTrailingComment(
-        enclosingNode.decorators[enclosingNode.decorators.length - 1],
-        comment
-      );
-    }
-    return true;
-  }
-  return false;
-}
-
-function handleMethodNameComments(text, enclosingNode, precedingNode, comment) {
-  // This is only needed for estree parsers (flow, typescript) to attach
-  // after a method name:
-  // obj = { fn /*comment*/() {} };
-  if (
-    enclosingNode &&
-    precedingNode &&
-    (enclosingNode.type === "Property" ||
-      enclosingNode.type === "MethodDefinition") &&
-    precedingNode.type === "Identifier" &&
-    enclosingNode.key === precedingNode &&
-    // special Property case: { key: /*comment*/(value) };
-    // comment should be attached to value instead of key
-    getNextNonSpaceNonCommentCharacter(text, precedingNode) !== ":"
-  ) {
-    addTrailingComment(precedingNode, comment);
-    return true;
-  }
-
-  // Print comments between decorators and class methods as a trailing comment
-  // on the decorator node instead of the method node
-  if (
-    precedingNode &&
-    enclosingNode &&
-    precedingNode.type === "Decorator" &&
-    (enclosingNode.type === "ClassMethod" ||
-      enclosingNode.type === "ClassProperty" ||
-      enclosingNode.type === "TSAbstractClassProperty" ||
-      enclosingNode.type === "TSAbstractMethodDefinition" ||
-      enclosingNode.type === "MethodDefinition")
-  ) {
-    addTrailingComment(precedingNode, comment);
-    return true;
-  }
-
-  return false;
-}
-
-function handleFunctionNameComments(
-  text,
-  enclosingNode,
-  precedingNode,
-  comment
-) {
-  if (getNextNonSpaceNonCommentCharacter(text, comment) !== "(") {
-    return false;
-  }
-
-  if (
-    precedingNode &&
-    enclosingNode &&
-    (enclosingNode.type === "FunctionDeclaration" ||
-      enclosingNode.type === "FunctionExpression" ||
-      enclosingNode.type === "ClassMethod" ||
-      enclosingNode.type === "MethodDefinition" ||
-      enclosingNode.type === "ObjectMethod")
-  ) {
-    addTrailingComment(precedingNode, comment);
-    return true;
-  }
-  return false;
-}
-
-function handleCommentAfterArrowParams(text, enclosingNode, comment) {
-  if (!(enclosingNode && enclosingNode.type === "ArrowFunctionExpression")) {
-    return false;
-  }
-
-  const index = getNextNonSpaceNonCommentCharacterIndex(text, comment);
-  if (text.substr(index, 2) === "=>") {
-    addDanglingComment(enclosingNode, comment);
-    return true;
-  }
-
-  return false;
-}
-
-function handleCommentInEmptyParens(text, enclosingNode, comment) {
-  if (getNextNonSpaceNonCommentCharacter(text, comment) !== ")") {
-    return false;
-  }
-
-  // Only add dangling comments to fix the case when no params are present,
-  // i.e. a function without any argument.
-  if (
-    enclosingNode &&
-    (((enclosingNode.type === "FunctionDeclaration" ||
-      enclosingNode.type === "FunctionExpression" ||
-      (enclosingNode.type === "ArrowFunctionExpression" &&
-        (enclosingNode.body.type !== "CallExpression" ||
-          enclosingNode.body.arguments.length === 0)) ||
-      enclosingNode.type === "ClassMethod" ||
-      enclosingNode.type === "ObjectMethod") &&
-      enclosingNode.params.length === 0) ||
-      (enclosingNode.type === "CallExpression" &&
-        enclosingNode.arguments.length === 0))
-  ) {
-    addDanglingComment(enclosingNode, comment);
-    return true;
-  }
-  if (
-    enclosingNode &&
-    (enclosingNode.type === "MethodDefinition" &&
-      enclosingNode.value.params.length === 0)
-  ) {
-    addDanglingComment(enclosingNode.value, comment);
-    return true;
-  }
-  return false;
-}
-
-function handleLastFunctionArgComments(
-  text,
-  precedingNode,
-  enclosingNode,
-  followingNode,
-  comment
-) {
-  // Type definitions functions
-  if (
-    precedingNode &&
-    precedingNode.type === "FunctionTypeParam" &&
-    enclosingNode &&
-    enclosingNode.type === "FunctionTypeAnnotation" &&
-    followingNode &&
-    followingNode.type !== "FunctionTypeParam"
-  ) {
-    addTrailingComment(precedingNode, comment);
-    return true;
-  }
-
-  // Real functions
-  if (
-    precedingNode &&
-    (precedingNode.type === "Identifier" ||
-      precedingNode.type === "AssignmentPattern") &&
-    enclosingNode &&
-    (enclosingNode.type === "ArrowFunctionExpression" ||
-      enclosingNode.type === "FunctionExpression" ||
-      enclosingNode.type === "FunctionDeclaration" ||
-      enclosingNode.type === "ObjectMethod" ||
-      enclosingNode.type === "ClassMethod") &&
-    getNextNonSpaceNonCommentCharacter(text, comment) === ")"
-  ) {
-    addTrailingComment(precedingNode, comment);
-    return true;
-  }
-  return false;
-}
-
-function handleImportSpecifierComments(enclosingNode, comment) {
-  if (enclosingNode && enclosingNode.type === "ImportSpecifier") {
-    addLeadingComment(enclosingNode, comment);
-    return true;
-  }
-  return false;
-}
-
-function handleLabeledStatementComments(enclosingNode, comment) {
-  if (enclosingNode && enclosingNode.type === "LabeledStatement") {
-    addLeadingComment(enclosingNode, comment);
-    return true;
-  }
-  return false;
-}
-
-function handleCallExpressionComments(precedingNode, enclosingNode, comment) {
-  if (
-    enclosingNode &&
-    enclosingNode.type === "CallExpression" &&
-    precedingNode &&
-    enclosingNode.callee === precedingNode &&
-    enclosingNode.arguments.length > 0
-  ) {
-    addLeadingComment(enclosingNode.arguments[0], comment);
-    return true;
-  }
-  return false;
-}
-
-function handleUnionTypeComments(
-  precedingNode,
-  enclosingNode,
-  followingNode,
-  comment
-) {
-  if (
-    enclosingNode &&
-    (enclosingNode.type === "UnionTypeAnnotation" ||
-      enclosingNode.type === "TSUnionType")
-  ) {
-    addTrailingComment(precedingNode, comment);
-    return true;
-  }
-  return false;
-}
-
-function handlePropertyComments(enclosingNode, comment) {
-  if (
-    enclosingNode &&
-    (enclosingNode.type === "Property" ||
-      enclosingNode.type === "ObjectProperty")
-  ) {
-    addLeadingComment(enclosingNode, comment);
-    return true;
-  }
-  return false;
-}
-
-function handleExportNamedDeclarationComments(enclosingNode, comment) {
-  if (enclosingNode && enclosingNode.type === "ExportNamedDeclaration") {
-    addLeadingComment(enclosingNode, comment);
-    return true;
-  }
-  return false;
-}
-
-function handleOnlyComments(enclosingNode, ast, comment, isLastComment) {
-  // With Flow the enclosingNode is undefined so use the AST instead.
-  if (ast && ast.body && ast.body.length === 0) {
-    if (isLastComment) {
-      addDanglingComment(ast, comment);
-    } else {
-      addLeadingComment(ast, comment);
-    }
-    return true;
-  } else if (
-    enclosingNode &&
-    enclosingNode.type === "Program" &&
-    enclosingNode.body.length === 0 &&
-    enclosingNode.directives &&
-    enclosingNode.directives.length === 0
-  ) {
-    if (isLastComment) {
-      addDanglingComment(enclosingNode, comment);
-    } else {
-      addLeadingComment(enclosingNode, comment);
-    }
-    return true;
-  }
-  return false;
-}
-
-function handleForComments(enclosingNode, precedingNode, comment) {
-  if (
-    enclosingNode &&
-    (enclosingNode.type === "ForInStatement" ||
-      enclosingNode.type === "ForOfStatement")
-  ) {
-    addLeadingComment(enclosingNode, comment);
-    return true;
-  }
-  return false;
-}
-
-function handleImportDeclarationComments(
-  text,
-  enclosingNode,
-  precedingNode,
-  comment
-) {
-  if (
-    precedingNode &&
-    enclosingNode &&
-    enclosingNode.type === "ImportDeclaration" &&
-    util.hasNewline(text, util.locEnd(comment))
-  ) {
-    addTrailingComment(precedingNode, comment);
-    return true;
-  }
-  return false;
-}
-
-function handleAssignmentPatternComments(enclosingNode, comment) {
-  if (enclosingNode && enclosingNode.type === "AssignmentPattern") {
-    addLeadingComment(enclosingNode, comment);
-    return true;
-  }
-  return false;
-}
-
-function handleTypeAliasComments(enclosingNode, followingNode, comment) {
-  if (enclosingNode && enclosingNode.type === "TypeAlias") {
-    addLeadingComment(enclosingNode, comment);
-    return true;
-  }
-  return false;
-}
-
-function handleVariableDeclaratorComments(
-  enclosingNode,
-  followingNode,
-  comment
-) {
-  if (
-    enclosingNode &&
-    enclosingNode.type === "VariableDeclarator" &&
-    followingNode &&
-    (followingNode.type === "ObjectExpression" ||
-      followingNode.type === "ArrayExpression")
-  ) {
-    addLeadingComment(followingNode, comment);
-    return true;
-  }
-  return false;
-}
-
 function printComment(commentPath, options) {
   const comment = commentPath.getValue();
   comment.printed = true;
   return options.printer.printComment(commentPath, options);
 }
 
-function findExpressionIndexForComment(quasis, comment) {
-  const startPos = locStart(comment) - 1;
+function findExpressionIndexForComment(quasis, comment, options) {
+  const startPos = options.locStart(comment) - 1;
 
   for (let i = 1; i < quasis.length; ++i) {
     if (startPos < getQuasiRange(quasis[i]).start) {
@@ -954,14 +371,17 @@ function printLeadingComment(commentPath, print, options) {
   if (!contents) {
     return "";
   }
-  const isBlock = util.isBlockComment(comment);
+  const isBlock =
+    options.printer.isBlockComment && options.printer.isBlockComment(comment);
 
   // Leading block comments should see if they need to stay on the
   // same line or not.
   if (isBlock) {
     return concat([
       contents,
-      util.hasNewline(options.originalText, locEnd(comment)) ? hardline : " "
+      privateUtil.hasNewline(options.originalText, options.locEnd(comment))
+        ? hardline
+        : " "
     ]);
   }
 
@@ -974,7 +394,8 @@ function printTrailingComment(commentPath, print, options) {
   if (!contents) {
     return "";
   }
-  const isBlock = util.isBlockComment(comment);
+  const isBlock =
+    options.printer.isBlockComment && options.printer.isBlockComment(comment);
 
   // We don't want the line to break
   // when the parentParentNode is a ClassDeclaration/-Expression
@@ -988,7 +409,7 @@ function printTrailingComment(commentPath, print, options) {
     parentParentNode.superClass === parentNode;
 
   if (
-    util.hasNewline(options.originalText, locStart(comment), {
+    privateUtil.hasNewline(options.originalText, options.locStart(comment), {
       backwards: true
     })
   ) {
@@ -1004,9 +425,10 @@ function printTrailingComment(commentPath, print, options) {
     // if this a comment on its own line; normal trailing comments are
     // always at the end of another expression.
 
-    const isLineBeforeEmpty = util.isPreviousLineEmpty(
+    const isLineBeforeEmpty = privateUtil.isPreviousLineEmpty(
       options.originalText,
-      comment
+      comment,
+      options.locStart
     );
 
     return lineSuffix(
@@ -1052,7 +474,7 @@ function printDanglingComments(path, options, sameIndent, filter) {
 
 function prependCursorPlaceholder(path, options, printed) {
   if (path.getNode() === options.cursorNode && path.getValue()) {
-    return concat([cursor, printed]);
+    return concat([cursor, printed, cursor]);
   }
   return printed;
 }
@@ -1082,7 +504,12 @@ function printComments(path, print, options, needsSemi) {
       leadingParts.push(contents);
 
       const text = options.originalText;
-      if (util.hasNewline(text, util.skipNewline(text, util.locEnd(comment)))) {
+      if (
+        privateUtil.hasNewline(
+          text,
+          privateUtil.skipNewline(text, options.locEnd(comment))
+        )
+      ) {
         leadingParts.push(hardline);
       }
     } else if (trailing) {
