@@ -4,8 +4,7 @@ const prettier = require("../../tests_config/require_prettier");
 const runPrettier = require("../runPrettier");
 const constant = require("../../src/cli/constant");
 const util = require("../../src/cli/util");
-const commonUtil = require("../../src/common/util");
-const getSupportInfo = require("../../src/common/support").getSupportInfo;
+const arrayify = require("../../src/utils/arrayify");
 
 describe("show version with --version", () => {
   runPrettier("cli/with-shebang", ["--version"]).test({
@@ -26,35 +25,55 @@ describe(`show detailed usage with --help l (alias)`, () => {
   });
 });
 
-commonUtil
-  .arrayify(
-    Object.assign(
-      {},
-      util.createDetailedOptionMap(
-        getSupportInfo(null, {
-          showDeprecated: true,
-          showUnreleased: true,
-          showInternal: true
-        }).options
-      ),
-      util.normalizeDetailedOptionMap(constant.options)
-    ),
-    "name"
-  )
-  .forEach(option => {
-    const optionNames = [
-      option.description ? option.name : null,
-      option.oppositeDescription ? `no-${option.name}` : null
-    ].filter(Boolean);
+describe(`show detailed usage with plugin options (automatic resolution)`, () => {
+  runPrettier("plugins/automatic", [
+    "--help",
+    "tab-width",
+    "--parser=bar",
+    `--plugin-search-dir=.`
+  ]).test({
+    status: 0
+  });
+});
 
-    optionNames.forEach(optionName => {
-      describe(`show detailed usage with --help ${optionName}`, () => {
-        runPrettier("cli", ["--help", optionName]).test({
-          status: 0
-        });
+describe(`show detailed usage with plugin options (manual resolution)`, () => {
+  runPrettier("cli", [
+    "--help",
+    "tab-width",
+    "--plugin=../plugins/automatic/node_modules/prettier-plugin-bar",
+    "--parser=bar"
+  ]).test({
+    status: 0
+  });
+});
+
+arrayify(
+  Object.assign(
+    {},
+    util.createDetailedOptionMap(
+      prettier.getSupportInfo(null, {
+        showDeprecated: true,
+        showUnreleased: true,
+        showInternal: true
+      }).options
+    ),
+    util.normalizeDetailedOptionMap(constant.options)
+  ),
+  "name"
+).forEach(option => {
+  const optionNames = [
+    option.description ? option.name : null,
+    option.oppositeDescription ? `no-${option.name}` : null
+  ].filter(Boolean);
+
+  optionNames.forEach(optionName => {
+    describe(`show detailed usage with --help ${optionName}`, () => {
+      runPrettier("cli", ["--help", optionName]).test({
+        status: 0
       });
     });
   });
+});
 
 describe("show warning with --help not-found", () => {
   runPrettier("cli", ["--help", "not-found"]).test({
@@ -76,6 +95,12 @@ describe("throw error with --write + --debug-check", () => {
 
 describe("throw error with --find-config-path + multiple files", () => {
   runPrettier("cli", ["--find-config-path", "abc.js", "def.js"]).test({
+    status: 1
+  });
+});
+
+describe("throw error with --file-info + multiple files", () => {
+  runPrettier("cli", ["--file-info", "abc.js", "def.js"]).test({
     status: 1
   });
 });
