@@ -1,3 +1,3 @@
 function js() {
-  console.log("js/file.js should have tab width 8");
+  console.log("js/file.js should have tab width 8 (1 if CLI)");
 }
