@@ -7,8 +7,7 @@ module.exports = {
     {
       name: "foo",
       parsers: ["foo-parser"],
-      extensions: [".foo"],
-      since: "1.0.0"
+      extensions: [".foo"]
     }
   ],
   parsers: {
